+++ dev-site-config/generatedComponentConfig.js
@@ -253,18 +253,20 @@ import ControlledRadioField from '../packages/terra-form-radio/examples/test-exa
 import HiddenLegend1 from '../packages/terra-form-radio/examples/test-examples/radio-field/HiddenLegend.example';
 import HideRequiredInvalidRadioField from '../packages/terra-form-radio/examples/test-examples/radio-field/HideRequiredInvalidRadioField.example';
 import OptionalRadioField from '../packages/terra-form-radio/examples/test-examples/radio-field/OptionalRadioField.example';
-import DefaultSelectOption from '../packages/terra-form-select/tests/nightwatch/select-option/DefaultSelectOption';
-import DisabledSelectOption from '../packages/terra-form-select/tests/nightwatch/select-option/DisabledSelectOption';
-import SelectOptionLongText from '../packages/terra-form-select/tests/nightwatch/select-option/SelectOptionLongText';
-import SelectOptionWithChildren from '../packages/terra-form-select/tests/nightwatch/select-option/SelectOptionWithChildren';
-import SelectedSelectOption from '../packages/terra-form-select/tests/nightwatch/select-option/SelectedSelectOption';
-import BlankSelect from '../packages/terra-form-select/tests/nightwatch/select/BlankSelect';
-import ControlledSelect from '../packages/terra-form-select/tests/nightwatch/select/ControlledSelect';
-import DefaultSelect from '../packages/terra-form-select/tests/nightwatch/select/DefaultSelect';
-import DisabledSelect from '../packages/terra-form-select/tests/nightwatch/select/DisabledSelect';
-import HiddenPlaceholderSelect from '../packages/terra-form-select/tests/nightwatch/select/HiddenPlaceholderSelect';
-import InvalidSelect from '../packages/terra-form-select/tests/nightwatch/select/InvalidSelect';
-import SelectLongText from '../packages/terra-form-select/tests/nightwatch/select/SelectLongText';
+import ControlledCombobox from '../packages/terra-form-select/examples/test-examples/ControlledCombobox.example';
+import ControlledDefault from '../packages/terra-form-select/examples/test-examples/ControlledDefault.example';
+import ControlledMultiple from '../packages/terra-form-select/examples/test-examples/ControlledMultiple.example';
+import ControlledSearch from '../packages/terra-form-select/examples/test-examples/ControlledSearch.example';
+import ControlledTag from '../packages/terra-form-select/examples/test-examples/ControlledTag.example';
+import Disabled from '../packages/terra-form-select/examples/test-examples/Disabled.example';
+import ManyOptions from '../packages/terra-form-select/examples/test-examples/ManyOptions.example';
+import OptGroup from '../packages/terra-form-select/examples/test-examples/OptGroup.example';
+import UncontrolledCombobox from '../packages/terra-form-select/examples/test-examples/UncontrolledCombobox.example';
+import UncontrolledDefault from '../packages/terra-form-select/examples/test-examples/UncontrolledDefault.example';
+import UncontrolledDefaultLongText from '../packages/terra-form-select/examples/test-examples/UncontrolledDefaultLongText.example';
+import UncontrolledMultiple from '../packages/terra-form-select/examples/test-examples/UncontrolledMultiple.example';
+import UncontrolledSearch from '../packages/terra-form-select/examples/test-examples/UncontrolledSearch.example';
+import UncontrolledTag from '../packages/terra-form-select/examples/test-examples/UncontrolledTag.example';
 import TextareaField from '../packages/terra-form-textarea/examples/test-examples/TextareaField.example';
 import AutoResizableTextarea from '../packages/terra-form-textarea/tests/nightwatch/AutoResizableTextarea';
 import AutoResizableTextareaMobileDocumentTouch from '../packages/terra-form-textarea/tests/nightwatch/AutoResizableTextareaMobileDocumentTouch';
@@ -294,7 +296,7 @@ import DefaultSelectField from '../packages/terra-form/tests/nightwatch/select-f
 import PopulatedSelectField from '../packages/terra-form/tests/nightwatch/select-field/PopulatedSelectField';
 import PopulatedSelectFieldControlled from '../packages/terra-form/tests/nightwatch/select-field/PopulatedSelectFieldControlled';
 import ChoicesPropSelect from '../packages/terra-form/tests/nightwatch/select/ChoicesPropSelect';
-import DefaultSelect1 from '../packages/terra-form/tests/nightwatch/select/DefaultSelect';
+import DefaultSelect from '../packages/terra-form/tests/nightwatch/select/DefaultSelect';
 import PopulatedSelect from '../packages/terra-form/tests/nightwatch/select/PopulatedSelect';
 import PopulatedSelectControlled from '../packages/terra-form/tests/nightwatch/select/PopulatedSelectControlled';
 import DefaultTextField from '../packages/terra-form/tests/nightwatch/text-field/DefaultTextField';
@@ -1604,76 +1606,74 @@ const componentConfig = {
     path: '/form-select',
     tests: [
       {
-        name: 'Select Option',
-        path: '/select-option',
-        tests: [
-          {
-            name: 'Default Select Option',
-            path: '/default-select-option',
-            component: DefaultSelectOption,
-          },
-          {
-            name: 'Disabled Select Option',
-            path: '/disabled-select-option',
-            component: DisabledSelectOption,
-          },
-          {
-            name: 'Select Option Long Text',
-            path: '/select-option-long-text',
-            component: SelectOptionLongText,
-          },
-          {
-            name: 'Select Option With Children',
-            path: '/select-option-with-children',
-            component: SelectOptionWithChildren,
-          },
-          {
-            name: 'Selected Select Option',
-            path: '/selected-select-option',
-            component: SelectedSelectOption,
-          },
-        ],
+        name: 'Controlled Combobox',
+        path: '/controlled-combobox',
+        component: ControlledCombobox,
       },
       {
-        name: 'Select',
-        path: '/select',
-        tests: [
-          {
-            name: 'Blank Select',
-            path: '/blank-select',
-            component: BlankSelect,
-          },
-          {
-            name: 'Controlled Select',
-            path: '/controlled-select',
-            component: ControlledSelect,
-          },
-          {
-            name: 'Default Select',
-            path: '/default-select',
-            component: DefaultSelect,
-          },
-          {
-            name: 'Disabled Select',
-            path: '/disabled-select',
-            component: DisabledSelect,
-          },
-          {
-            name: 'Hidden Placeholder Select',
-            path: '/hidden-placeholder-select',
-            component: HiddenPlaceholderSelect,
-          },
-          {
-            name: 'Invalid Select',
-            path: '/invalid-select',
-            component: InvalidSelect,
-          },
-          {
-            name: 'Select Long Text',
-            path: '/select-long-text',
-            component: SelectLongText,
-          },
-        ],
+        name: 'Controlled Default',
+        path: '/controlled-default',
+        component: ControlledDefault,
+      },
+      {
+        name: 'Controlled Multiple',
+        path: '/controlled-multiple',
+        component: ControlledMultiple,
+      },
+      {
+        name: 'Controlled Search',
+        path: '/controlled-search',
+        component: ControlledSearch,
+      },
+      {
+        name: 'Controlled Tag',
+        path: '/controlled-tag',
+        component: ControlledTag,
+      },
+      {
+        name: 'Disabled',
+        path: '/disabled',
+        component: Disabled,
+      },
+      {
+        name: 'Many Options',
+        path: '/many-options',
+        component: ManyOptions,
+      },
+      {
+        name: 'Opt Group',
+        path: '/opt-group',
+        component: OptGroup,
+      },
+      {
+        name: 'Uncontrolled Combobox',
+        path: '/uncontrolled-combobox',
+        component: UncontrolledCombobox,
+      },
+      {
+        name: 'Uncontrolled Default',
+        path: '/uncontrolled-default',
+        component: UncontrolledDefault,
+      },
+      {
+        name: 'Uncontrolled Default Long Text',
+        path: '/uncontrolled-default-long-text',
+        component: UncontrolledDefaultLongText,
+      },
+      {
+        name: 'Uncontrolled Multiple',
+        path: '/uncontrolled-multiple',
+        component: UncontrolledMultiple,
+      },
+      {
+        name: 'Uncontrolled Search',
+        path: '/uncontrolled-search',
+        component: UncontrolledSearch,
+      },
+      {
+        name: 'Uncontrolled Tag',
+        path: '/uncontrolled-tag',
+        component: UncontrolledTag,
       },
     ],
   },
@@ -1875,7 +1875,7 @@ const componentConfig = {
           {
             name: 'Default Select',
             path: '/default-select',
-            component: DefaultSelect1,
+            component: DefaultSelect,
           },
           {
             name: 'Populated Select',